+++ plugin/src/integrationTest/kotlin/com/vanniktech/maven/publish/ProjectSpecRunner.kt
@@ -20,7 +20,7 @@ fun ProjectSpec.run(fixtures: Path, temp: Path, options: TestOptions): ProjectRe
 
   val task = ":publishAllPublicationsToTestFolderRepository"
   val arguments = mutableListOf(task, "--stacktrace")
-  if (supportsConfigCaching(plugins)) {
+  if (supportsConfigCaching()) {
     arguments.add("--configuration-cache")
   }
 
@@ -40,9 +40,10 @@ fun ProjectSpec.run(fixtures: Path, temp: Path, options: TestOptions): ProjectRe
   )
 }
 
-private fun supportsConfigCaching(plugins: List<PluginSpec>): Boolean {
-  // TODO https://github.com/Kotlin/dokka/issues/2231
-  return !plugins.any { it.id == dokkaPlugin.id }
+private fun ProjectSpec.supportsConfigCaching(): Boolean {
+  // TODO can always return true when dropping support for dokka in v1 mode
+  //  to simplify the test set up this assumes that version 2.x always runs in v2 mode
+  return plugins.none { it.id == dokkaPlugin.id && it.version!!.startsWith("1.") }
 }
 
 private fun ProjectSpec.writeBuildFile(path: Path, repo: Path, options: TestOptions) {
@@ -187,6 +188,8 @@ private fun ProjectSpec.writeGradleProperties(path: Path, options: TestOptions) 
       appendLine("org.gradle.vfs.watch=false")
       appendLine("kotlin.compiler.execution.strategy=in-process")
       appendLine("kotlin.mpp.androidSourceSetLayoutVersion1.nowarn=true")
+      appendLine("org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled")
+      appendLine("org.jetbrains.dokka.experimental.gradle.pluginMode.noWarn=true")
       appendLine()
 
       if (options.config == TestOptions.Config.PROPERTIES) {