+++ plugin/src/main/kotlin/com/vanniktech/maven/publish/MavenPublishBaseExtension.kt
@@ -406,12 +406,19 @@ abstract class MavenPublishBaseExtension @Inject constructor(
   private fun defaultJavaDocOption(javadocJar: Boolean, plainJavadocSupported: Boolean): JavadocJar {
     return if (!javadocJar) {
       JavadocJar.None()
-    } else if (project.plugins.hasPlugin("org.jetbrains.dokka") || project.plugins.hasPlugin("org.jetbrains.dokka-android")) {
-      val dokkaTask = project.provider {
-        val tasks = project.tasks.withType(DokkaTask::class.java)
-        tasks.singleOrNull()?.name ?: "dokkaHtml"
+    } else if (project.plugins.hasPlugin("org.jetbrains.dokka-javadoc")) {
+      JavadocJar.Dokka("dokkaGeneratePublicationJavadoc")
+    } else if (project.plugins.hasPlugin("org.jetbrains.dokka")) {
+      // only dokka v2 has an extension
+      if (project.extensions.findByName("dokka") != null) {
+        JavadocJar.Dokka("dokkaGeneratePublicationHtml")
+      } else {
+        val dokkaTask = project.provider {
+          val tasks = project.tasks.withType(DokkaTask::class.java)
+          tasks.singleOrNull()?.name ?: "dokkaHtml"
+        }
+        JavadocJar.Dokka(dokkaTask)
       }
-      JavadocJar.Dokka(dokkaTask)
     } else if (plainJavadocSupported) {
       project.tasks.withType(Javadoc::class.java).configureEach {
         val options = it.options as StandardJavadocDocletOptions